refactor: comply with Ludwig convention with GO_SYMBOL/END_SYMBOL

+++ ludwig/models/ecd.py
@@ -71,7 +71,7 @@ class ECD(tf.keras.Model):
             # use presence or absence of targets to signal training or prediction
             if targets is not None:
                 # doing training
-                target_to_use = targets[output_feature_name]
+                target_to_use = tf.cast(targets[output_feature_name], dtype=tf.int32)
             else:
                 # doing prediction
                 target_to_use = None
